+ new buffer test cases

+++ Engine/graphics/encoder.cpp
@@ -155,12 +155,12 @@ void Encoder<CommandBuffer>::implEndRenderPass() {
     }
   }
 
-void Encoder<CommandBuffer>::copy(const Attachment& src, size_t mip, StorageBuffer& dest, size_t offset) {
+void Encoder<CommandBuffer>::copy(const Attachment& src, uint32_t mip, StorageBuffer& dest, size_t offset) {
   uint32_t w = src.w(), h = src.h();
   impl->copy(*dest.impl.impl.handler,TextureLayout::Sampler,w,h,mip,*textureCast(src).impl.handler,offset);
   }
 
-void Encoder<CommandBuffer>::copy(const Texture2d& src, size_t mip, StorageBuffer& dest, size_t offset) {
+void Encoder<CommandBuffer>::copy(const Texture2d& src, uint32_t mip, StorageBuffer& dest, size_t offset) {
   uint32_t w = src.w(), h = src.h();
   impl->copy(*dest.impl.impl.handler,TextureLayout::Sampler,w,h,mip,*src.impl.handler,offset);
   }
