+++ Engine/graphics/encoder.h
@@ -71,8 +71,8 @@ class Encoder<Tempest::CommandBuffer> {
 
     void dispatch(size_t x, size_t y, size_t z);
 
-    void copy(const Attachment& src, size_t mip, StorageBuffer& dest, size_t offset);
-    void copy(const Texture2d&  src, size_t mip, StorageBuffer& dest, size_t offset);
+    void copy(const Attachment& src, uint32_t mip, StorageBuffer& dest, size_t offset);
+    void copy(const Texture2d&  src, uint32_t mip, StorageBuffer& dest, size_t offset);
 
     void generateMipmaps(Attachment& tex);
